notifications/requirements.txt
dnspython==2.7.0
motor==3.7.0
pika==1.3.2
pymongo==4.11.3
python-dotenv==1.1.0
